requirements.txt
alembic==1.14.1
altair==5.3.0
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==23.2.0
audioread==3.0.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
celery==5.4.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1734380492396/work/certifi
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
checksumdir==1.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.2.1
curl_cffi==0.7.0b4
cycler==0.12.1
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1737269753336/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
duckdb==0.10.2
duckduckgo_search==5.3.1
email_validator==2.2.0
exa-py==1.0.9
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
filelock==3.16.1
Flask==3.0.3
Flask-Excel==0.0.7
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.51.0
frozendict==2.4.4
fsspec==2024.12.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-aiplatform==1.77.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.27.1
idna==3.7
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources==6.4.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1734788142186/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver==1.4.5
kombu==5.4.0
lancedb==0.19.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
lml==0.1.0
lxml==5.2.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multitasking==0.0.11
narwhals==1.22.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.0.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.59.8
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
passlib==1.7.4
peewee==3.17.5
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pgvector==0.2.5
phidata==2.7.10
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==10.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
pooch==1.8.2
primp==0.10.1
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.29.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1735327328223/work
psycopg==3.1.18
psycopg-binary==3.1.18
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==16.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pydeck==0.9.1
pydub==0.25.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.23.0
pyparsing==3.1.2
pypdf==4.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1728642222605/work
redis==5.0.8
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.13.0
soupsieve==2.5
soxr==0.5.0.post1
SQLAlchemy==2.0.31
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
streamlit==1.34.0
sympy==1.13.1
tantivy==0.22.0
tenacity==8.3.0
texttable==1.7.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.5.1
torchaudio==2.5.1
tornado==6.4
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.48.0
triton==3.1.0
typer==0.12.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
yfinance==0.2.38
yt-dlp==2025.1.26
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
